fix local dependency problem

+++ google-cloud-pubsub/src/test/java/com/google/cloud/pubsub/PublisherImplTest.java
@@ -407,10 +407,7 @@ public class PublisherImplTest {
     }
     try {
       builder.setBundlingSettings(
-          Publisher.DEFAULT_BUNDLING_SETTINGS
-              .toBuilder()
-              .setRequestByteThreshold((Long) null)
-              .build());
+          Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setRequestByteThreshold(null).build());
       fail("Should have thrown an NullPointerException");
     } catch (NullPointerException expected) {
       // Expected
@@ -454,10 +451,7 @@ public class PublisherImplTest {
         Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setElementCountThreshold(1).build());
     try {
       builder.setBundlingSettings(
-          Publisher.DEFAULT_BUNDLING_SETTINGS
-              .toBuilder()
-              .setElementCountThreshold((Long) null)
-              .build());
+          Publisher.DEFAULT_BUNDLING_SETTINGS.toBuilder().setElementCountThreshold(null).build());
       fail("Should have thrown an NullPointerException");
     } catch (NullPointerException expected) {
       // Expected
